% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.sgpv}
\alias{coef.sgpv}
\title{coef S3 method}
\usage{
\method{coef}{sgpv}(object, ...)
}
\arguments{
\item{object}{An \code{sgpv} object}

\item{...}{Other \code{coef} arguments}
}
\value{
Coefficients in the OLS model
}
\description{
Extract coefficients from the model fit
}
