% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.candidate}
\alias{get.candidate}
\title{\code{get.candidate}: Get candidate set}
\usage{
get.candidate(xs, ys, family)
}
\arguments{
\item{xs}{Standardized independent variables}

\item{ys}{Standardized dependent variable}

\item{family}{A description of the error distribution and link function to be
used in the model. It can take the value of \verb{\code{gaussian}}, \verb{\code{binomial}},
\verb{\code{poisson}}, and \verb{\code{cox}}.}
}
\value{
A list of following components:
\describe{
\item{candidate.index}{A vector of indices of selected variables in the candidate set}
\item{lambda}{The \code{lambda} selected by generalized information criterion}
}
}
\description{
Get the indices of the candidate set in the first stage
}
