% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.coef}
\alias{get.coef}
\title{\code{get.coef}: Get coefficients at each \code{lambda}}
\usage{
get.coef(xs, ys, lambda, lasso, family)
}
\arguments{
\item{xs}{Standardized design matrix}

\item{ys}{Standardized outcome}

\item{lambda}{\code{lambda} in the lasso}

\item{lasso}{An \code{glmnet} object}

\item{family}{A description of the error distribution and link function to be
used in the model. It can take the value of \verb{\code{gaussian}}, \verb{\code{binomial}},
\verb{\code{poisson}}, and \verb{\code{cox}}.}
}
\value{
A vector that contains the point estimates, confidence intervals and the null bound
}
\description{
Get the coefficients and confidence intervals from regression at each \code{lambda}
as well as the null bound in SGPVs
}
