% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.var}
\alias{get.var}
\title{Get indices}
\usage{
get.var(candidate.index, xs, ys, family)
}
\arguments{
\item{candidate.index}{Indices of the candidate set}

\item{xs}{Standardized independent variables}

\item{ys}{Standardized dependent variable}

\item{family}{A description of the error distribution and link function to be
used in the model. It can take the value of \verb{\code{gaussian}}, \verb{\code{binomial}},
\verb{\code{poisson}}, and \verb{\code{cox}}.}
}
\value{
Indices of variables selected
}
\description{
Get the indices of the variables selected by the algorithm
}
