% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gvif}
\alias{gvif}
\title{\code{gvif}: Get GVIF for each variable}
\usage{
gvif(mod, family)
}
\arguments{
\item{mod}{A model object with at least two explanatory variables}

\item{family}{A description of the error distribution and link function to be
used in the model. It can take the value of \verb{\code{gaussian}}, \verb{\code{binomial}},
\verb{\code{poisson}}, and \verb{\code{cox}}.}
}
\value{
A vector of GVIF for each variable in the model
}
\description{
Get generalized variance inflation factor (GVIF) for each variable.
See Fox (1992) \doi{10.1080/01621459.1992.10475190} for more details on how to calculate GVIF.
}
