% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{one.time.size}
\alias{one.time.size}
\title{\code{one.time.size}: utility function for \code{which.sgpv} function}
\usage{
one.time.size(x, y, family)
}
\arguments{
\item{x}{Input X}

\item{y}{Input Y}

\item{family}{A description of the error distribution and link function to be
used in the model. It can take the value of \verb{\code{gaussian}}, \verb{\code{binomial}},
\verb{\code{poisson}}, and \verb{\code{cox}}. Default is \verb{\code{gaussian}}}
}
\value{
A list of following components:
\describe{
\item{model}{The indices of selected variables}
\item{size}{The size of the model}
}
}
\description{
Return the selected model as well as the model size in each iteration
}
