% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sgpv}
\alias{plot.sgpv}
\title{plot S3 method}
\usage{
\method{plot}{sgpv}(x, lpv = c(1, 3), lambda.max = NULL, ...)
}
\arguments{
\item{x}{An \code{sgpv} object}

\item{lpv}{Lines per variable. It can take the value of 1 meaning that only the
bound that is closest to the null will be plotted, or the value of 3 meaning that
point estimates as well as 95\% confidence interval will be plotted}

\item{lambda.max}{The maximum lambda on the plot. Default is \code{NULL}.}

\item{...}{Other \code{plot} arguments}
}
\description{
Plot the fully relaxed lasso solution path on
the standardized scale and the final variable selection results
}
