% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sgpv}
\alias{predict.sgpv}
\title{predict S3 method}
\usage{
\method{predict}{sgpv}(object, newx, ...)
}
\arguments{
\item{object}{An \code{sgpv} objectect}

\item{newx}{Prediction data set}

\item{...}{Other \code{predict} arguments}
}
\value{
Predicted values
}
\description{
Prediction using the fitted model
}
