% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{print.sgpv}
\alias{print.sgpv}
\title{print S3 method}
\usage{
\method{print}{sgpv}(x, ...)
}
\arguments{
\item{x}{An \code{sgpv} object}

\item{...}{Other \code{print} arguments}
}
\value{
Variable selection results
}
\description{
print S3 method
}
