% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{pro.sgpv}
\alias{pro.sgpv}
\title{pro.sgpv function}
\usage{
pro.sgpv(x, y, stage = c(1, 2))
}
\arguments{
\item{x}{Independent variables, can be a \code{matrix} or a \code{data.frame}}

\item{y}{Dependent variable, can be a \code{vector} or a column from a \code{data.frame}}

\item{stage}{Algorithm indicator. 1 denotes the one-stage algorithm and 2 denotes the two-stage algorithm}
}
\value{
A list of following components:
\describe{
\item{out.sgpv}{A vector of indices of selected variables}
\item{lambda}{Cross-validated lambda in the two-stage algorithm. \code{NULL} for the one-stage algorithm}
}
}
\description{
This function outputs indices of selected variables
from either one-stage algorithm or two-stage algorithm and lambda.1se
from the cross-validated lasso
}
\examples{
library(ProSGPV)

x = t.housing[,-ncol(t.housing)]
y = t.housing$V9
out.sgpv.1 <- pro.sgpv(x = x, y = y, stage = 1)

}
