% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.sgpv}
\alias{which.sgpv}
\title{Find the most frequent model selected by the two-stage ProSGPV}
\usage{
which.sgpv(num.sim = 1000, object)
}
\arguments{
\item{num.sim}{Number of repetitions. Default is 1000}

\item{object}{An \code{sgpv} object}
}
\value{
A list of following components:
\describe{
\item{var.index}{The indices of selected variables in the most frequent model}
\item{var.label}{Labels of selected variables in the most frequent model}
\item{random.seed}{A random seed to reproduce the selection result}
\item{hist.plot}{A histogram of model size over repetitions}
}
}
\description{
Return the selected variables in the most frequent model selected by ProSGPV,
the random seed to produce the result, and a histogram of model size
}
