% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.coef}
\alias{get.coef}
\title{Get coefficients at each \code{lambda}}
\usage{
get.coef(xs, ys, lambda, lasso)
}
\arguments{
\item{xs}{Standardized design matrix}

\item{ys}{Standardized outcome}

\item{lambda}{\code{lambda} in the lasso}

\item{lasso}{An \code{glmnet} object}
}
\value{
A vector that contains the point estimates, confidence intervals and the null bound
}
\description{
Get the coefficients and confidence intervals from regression at each \code{lambda}
as well as the null bound in SGPVs
}
