% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.var}
\alias{get.var}
\title{Get indices}
\usage{
get.var(candidate.index, xs, ys)
}
\arguments{
\item{candidate.index}{Indices of the candidate set}

\item{xs}{Standardized independent variables}

\item{ys}{Standardized dependent variable}
}
\value{
Indices of variables selected
}
\description{
Get the indices of the variables selected by the algorithm
}
