% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{print.sgpv}
\alias{print.sgpv}
\title{\code{print.sgpv}: Print variable selection results}
\usage{
\method{print}{sgpv}(x, ...)
}
\arguments{
\item{x}{An \code{sgpv} object}

\item{...}{Other \code{print} arguments}
}
\value{
Variable selection results
}
\description{
S3 method \code{print} for an S3 object of class \code{sgpv}
}
\examples{

# prepare the data
x <- t.housing[, -ncol(t.housing)]
y <- t.housing$V9

# run one-stage algorithm
out.sgpv.1 <- pro.sgpv(x = x, y = y, stage = 1)

out.sgpv.1
}
