% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{pro.sgpv}
\alias{pro.sgpv}
\title{pro.sgpv function}
\usage{
pro.sgpv(x, y, stage = c(1, 2))
}
\arguments{
\item{x}{Independent variables, can be a \code{matrix} or a \code{data.frame}}

\item{y}{Dependent variable, can be a \code{vector} or a column from a \code{data.frame}}

\item{stage}{Algorithm indicator. 1 denotes the one-stage algorithm and 2 denotes the two-stage algorithm}
}
\value{
A list of following components:
\describe{
\item{var.index}{A vector of indices of selected variables}
\item{var.label}{A vector of labels of selected variables}
\item{lambda}{Cross-validated lambda in the two-stage algorithm. \code{NULL} for the one-stage algorithm}
\item{x}{Input data \code{x}}
\item{y}{Input data \code{y}}
}
}
\description{
This function outputs the variable selection results
from either one-stage algorithm or two-stage algorithm.
}
\examples{
library(ProSGPV)

x = t.housing[,-ncol(t.housing)]
y = t.housing$V9

# one-stage algorithm
out.sgpv.1 <- pro.sgpv(x = x, y = y, stage = 1)

# check variable selection results
out.sgpv.1

# extract OLS estimates
coef(out.sgpv.1)

# get prediction from the model
predict(out.sgpv.1)

# two-stage algorithm
out.sgpv.2 <- pro.sgpv(x = x, y = y, stage = 2)

# plot the fully relaxed lasso solution path and final solution
# plot(out.sgpv.2)

}
