% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{summary.sgpv}
\alias{summary.sgpv}
\title{summary S3 method}
\usage{
\method{summary}{sgpv}(object, ...)
}
\arguments{
\item{object}{An \code{sgpv} object}

\item{...}{Other arguments}
}
\value{
Summary of an OLS model
}
\description{
Provide summary of the OLS model on the selected variables
}
