% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro.sgpv.R
\name{summary.sgpv}
\alias{summary.sgpv}
\title{\code{summary.sgpv}: Summary of the final model}
\usage{
\method{summary}{sgpv}(object, ...)
}
\arguments{
\item{object}{An \code{sgpv} object}

\item{...}{Other arguments}
}
\value{
Summary of a model
}
\description{
S3 method \code{summary} for an S3 object of class \code{sgpv}
}
\examples{

# prepare the data
x <- t.housing[, -ncol(t.housing)]
y <- t.housing$V9

# run one-stage algorithm
out.sgpv <- pro.sgpv(x = x, y = y)

# get regression summary
summary(out.sgpv)
}
