% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which.sgpv}
\alias{which.sgpv}
\title{which.sgpv: Find the most frequent model selected by the two-stage ProSGPV}
\usage{
which.sgpv(object, num.sim = 100)
}
\arguments{
\item{object}{An \code{sgpv} object}

\item{num.sim}{Number of repetitions. Default is 1000}
}
\value{
A list of following components:
\describe{
\item{var.index}{The indices of selected variables in the most frequent model}
\item{var.label}{Labels of selected variables in the most frequent model}
\item{random.seed}{A random seed to reproduce the selection result}
\item{models}{A list of unqiue models sorted by frequency}
}
}
\description{
Return the selected variables in the most frequent model selected by ProSGPV,
the random seed to produce the result, and a density plot of model size. Examples
can be found at https://github.com/zuoyi93/ProSGPV/tree/master/vignettes.
}
